Display language announcement in language overview
Part of #4178.
